Widget link to target="_parent" fix for TFS
There is a problem with the way the widget is hosted: when you click the widget it should open a new tab with the Server Task. The problem is that the `origin` header is set to `origin: null` in all the modern browsers, and this causes the browser to fail the CORS request to Octopus. The Octopus Server deliberately prevents CORS headers when the `origin: null` since this is generally set when hosting from the file system - and this is not an expected scenario.

The other interesting point is that the `origin: null` header is still set for the entire life of the tab - even setting the URL manually or force reloading doesn't fix it.

We've changed this link to `target="_parent"` which works in VSTS and TFS. The only downside is that the user needs to `CTRL-Click` / `CMD-Click` the link to open it in a new tab.

Fixes https://github.com/OctopusDeploy/Issues/issues/3491